Normalization of Epoch 06/14/15
The code from my last commit was expanded to work with 3 gratings. I have selected ranges which represent the continuum, points were made, a slope line fit to the points, then the data was normalized to the slope line. For each grating. I have taken out the code for the PLOTS and have placed it in "Normalization Plots.py" Copy and paste into code as needed.