File Name Changes and New Epochs
Instead of making one program to run 4 sets of data, it is better to make 1 program run 1 set of data, then copy and paste it to work with as many sets of data as you need. I renamed "PointSelection and Interpolation" to "Epoch_06_14_15_Normalization" then copy and pasted it 3 times, one for each epoch. The pertinant data was then imported by simply changing the data import file names on line 50-55. The files were then renamed to represent the relevant epoch. Now all one needs to do in order to normalize data to a first degree polynomial is import the files, then input new ranges for polyfit points in lines 106-122.