Fixed Weird Data Dip
The data dip was caused by fringe cases in grating intersections. I only accounted for both of the flux values being zero. However the error was due to one grating having a zero flux value and another having a non zero value. Additionally, the same fringe cases were casing false apsorptions in the error spectra, which is now fixed. I also added a new function called "remove_small" which removes all values in a list smaller than a certain threshold, this cuts off the small noise at the end of each grating. This is a more elegant solution than choosing regions to chop off because now one does not need to search for the wavelength to cut off. Rather just adjust the threshold in the "remove_small" function untill the data is correct.