Fixed more weird absorption dips
In the cut gratings section, the code replaces values outside a specified index with 0. However, I forgot to change the error values of the flux to  1. Meaning that for ONE point and one point only during each grating intersection the code saw a flux value with normal error and 0 flux. Causing the weighted average the skew heaviliy forming the absorption dip. This is now fixed.